Move the define of aligned_vector to an independent file.
